+++ api/client/install.go
@@ -3,61 +3,10 @@ package client
 import (
 	"bytes"
 	"encoding/json"
-	"fmt"
-	"io"
 	"net/http"
 
 	"github.com/replicatedhq/embedded-cluster/api/types"
 )
-
-type APIError struct {
-	StatusCode int    `json:"status_code"`
-	Message    string `json:"message"`
-}
-
-func (e *APIError) Error() string {
-	return fmt.Sprintf("status=%d, message=%q", e.StatusCode, e.Message)
-}
-
-var defaultHTTPClient = &http.Client{
-	Transport: &http.Transport{
-		Proxy: nil, // This is a local client so no proxy is needed
-	},
-}
-
-type Client interface {
-	GetInstall() (*types.Install, error)
-	SetInstallConfig(config types.InstallationConfig) (*types.Install, error)
-	SetInstallStatus(status types.InstallationStatus) (*types.Install, error)
-}
-
-type client struct {
-	apiURL     string
-	httpClient *http.Client
-}
-
-type ClientOption func(*client)
-
-func WithHTTPClient(httpClient *http.Client) ClientOption {
-	return func(c *client) {
-		c.httpClient = httpClient
-	}
-}
-
-func New(apiURL string, opts ...ClientOption) Client {
-	c := &client{
-		apiURL: apiURL + "/install",
-	}
-	for _, opt := range opts {
-		opt(c)
-	}
-
-	if c.httpClient == nil {
-		c.httpClient = defaultHTTPClient
-	}
-
-	return c
-}
 
 func (c *client) GetInstall() (*types.Install, error) {
 	req, err := http.NewRequest("GET", c.apiURL, nil)
@@ -65,6 +14,7 @@ func (c *client) GetInstall() (*types.Install, error) {
 		return nil, err
 	}
 	req.Header.Set("Content-Type", "application/json")
+	req.Header.Set("Authorization", c.token)
 
 	resp, err := c.httpClient.Do(req)
 	if err != nil {
@@ -96,6 +46,7 @@ func (c *client) SetInstallConfig(config types.InstallationConfig) (*types.Insta
 		return nil, err
 	}
 	req.Header.Set("Content-Type", "application/json")
+	req.Header.Set("Authorization", c.token)
 
 	resp, err := c.httpClient.Do(req)
 	if err != nil {
@@ -122,11 +73,12 @@ func (c *client) SetInstallStatus(status types.InstallationStatus) (*types.Insta
 		return nil, err
 	}
 
-	req, err := http.NewRequest("POST", c.apiURL+"/api/install/phase/set-status", bytes.NewBuffer(b))
+	req, err := http.NewRequest("POST", c.apiURL+"/api/install/set-status", bytes.NewBuffer(b))
 	if err != nil {
 		return nil, err
 	}
 	req.Header.Set("Content-Type", "application/json")
+	req.Header.Set("Authorization", c.token)
 
 	resp, err := c.httpClient.Do(req)
 	if err != nil {
@@ -145,17 +97,4 @@ func (c *client) SetInstallStatus(status types.InstallationStatus) (*types.Insta
 	}
 
 	return &install, nil
-}
-
-func errorFromResponse(resp *http.Response) error {
-	body, err := io.ReadAll(resp.Body)
-	if err != nil {
-		return fmt.Errorf("unexpected response: status=%d", resp.StatusCode)
-	}
-	var apiError APIError
-	err = json.Unmarshal(body, &apiError)
-	if err != nil {
-		return fmt.Errorf("unexpected response: status=%d, body=%q", resp.StatusCode, string(body))
-	}
-	return &apiError
 }
